Update README.md
Feel free to reject or accept my changes, all based on my professional obsession with clean English. I think all of the changes are worthy and clarifying, but I'm biased.

Here are my rationale list notes for your consideration:

1) In general, I try to reduce Overcapitalization. It's a real estate-wide problem, heh.

2) I think it's important to spell out organizational acronyms like IANA, IETF and FIPS. You'll see that I spell them out only at their first mention. I didn't bother for API, URL or REPL.

3) In general, "sub" words are not hyphenated or two words, so I don't think we should do PascalCase bicapitalization for field names like "SubCounty" or "SubParcel."

4) We've been stating it as UPI version 2.0, UPI Version 2.0, UPI 2.0 or UPI v2 in call notes, but not UPIv2.

5) I tend to remove most contractions from instructional documentation.

You'll see other style changes to tighten language, use commas for independent clauses and remove commas for dependent clauses. Boring stuff.

I don't see the "subcountry" concept in this documentation, not to be confused with "subcounty." I don't know offhand if that needs to be added here or later.